break out parameter initialization to own module
