+++ src/calibration/bundle_adjustment/bundle_adjust_functions.py
@@ -6,110 +6,21 @@ LOG_FILE = r"log\bundle_adjust_functions.log"
 LOG_FORMAT = " %(levelname)-8s [%(filename)s:%(lineno)d] %(message)s"
 
 logging.basicConfig(filename=LOG_FILE, filemode="w", format=LOG_FORMAT, level=LOG_LEVEL)
-from pathlib import Path
 
-import sys
+from pathlib import Path
 import cv2
 import numpy as np
-import pandas as pd
 
 from scipy.optimize import least_squares
 from scipy.sparse import lil_matrix
 import time
 
 from src.cameras.camera_array import CameraArray, CameraArrayBuilder
+from src.calibration.bundle_adjustment.get_init_params import get_2d_3d_points, get_camera_params
+
 
 # note I'm changing this from 9 as I'm reducing the values being solved for
 CAMERA_PARAM_COUNT = 6
-
-
-def get_camera_params(camera_array):
-    """for each camera build the CAMERA_PARAM_COUNT element parameter index
-    camera_params with shape (n_cameras, CAMERA_PARAM_COUNT)
-    contains initial estimates of parameters for all cameras.
-    First 3 components in each row form a rotation vector (https://en.wikipedia.org/wiki/Rodrigues%27_rotation_formula),
-    next 3 components form a translation vector
-    """
-    camera_params = None
-    for port, cam in camera_array.cameras.items():
-        port_param = cam.to_vector()
-        if camera_params is None:
-            camera_params = port_param
-        else:
-            camera_params = np.vstack([camera_params, port_param])
-
-    return camera_params
-
-
-def get_points_2d_df(points_csv_path):
-    points_df = pd.read_csv(points_csv_path)
-
-    points_2d_port_A = points_df[
-        ["port_A", "bundle", "id", "x_A_raw", "y_A_raw"]
-    ].rename(
-        columns={
-            "port_A": "camera",
-            "bundle": "sync_index",
-            "id": "point_id",
-            "x_A_raw": "x_2d",
-            "y_A_raw": "y_2d",
-        }
-    )
-
-    points_2d_port_B = points_df[
-        ["port_B", "bundle", "id", "x_B_raw", "y_B_raw"]
-    ].rename(
-        columns={
-            "port_B": "camera",
-            "bundle": "sync_index",
-            "id": "point_id",
-            "x_B_raw": "x_2d",
-            "y_B_raw": "y_2d",
-        }
-    )
-
-    points_2d_df = (
-        pd.concat([points_2d_port_A, points_2d_port_B])
-        .drop_duplicates()
-        .sort_values(["sync_index", "point_id", "camera"])
-        .rename(columns={"sync_index": "bundle"})
-    )
-    return points_2d_df
-
-
-# get 3d points with indices to merge back into points_2d
-def get_points_3d_df(points_csv_path):
-    points_df = pd.read_csv(points_csv_path)
-    points_3d_df = (
-        points_df[["bundle", "id", "pair", "x_pos", "y_pos", "z_pos"]]
-        .sort_values(["bundle", "id"])
-        .groupby(["bundle", "id"])
-        .agg({"x_pos": "mean", "y_pos": "mean", "z_pos": "mean", "pair": "size"})
-        .rename(
-            columns={"pair": "count", "x_pos": "x_3d", "y_pos": "y_3d", "z_pos": "z_3d"}
-        )
-        .reset_index()
-        .reset_index()
-        .rename(columns={"index": "index_3d", "id": "point_id"})
-    )
-    return points_3d_df
-
-
-def get_bundle_adjust_params(points_2d_df: pd.DataFrame, points_3d_df: pd.DataFrame):
-
-    merged_point_data = (
-        points_2d_df.merge(points_3d_df, how="left", on=["bundle", "point_id"])
-        .sort_values(["camera", "bundle", "point_id"])
-        .dropna()
-    )
-
-    camera_indices = np.array(merged_point_data["camera"], dtype=np.int64)
-    point_indices = np.array(merged_point_data["index_3d"], dtype=np.int64)
-    points_2d = np.array(merged_point_data[["x_2d", "y_2d"]])
-    points_3d = np.array(points_3d_df[["x_3d", "y_3d", "z_3d"]])
-    n_points = points_3d.shape[0]
-
-    return camera_indices, point_indices, points_2d, points_3d, n_points
 
 
 def reprojection_error(
@@ -120,7 +31,7 @@ def reprojection_error(
     camera_params = params[: n_cameras * CAMERA_PARAM_COUNT].reshape(
         (n_cameras, CAMERA_PARAM_COUNT)
     )
-    
+
     # reshape the 3d points from vector to nx3
     points_3d = params[n_cameras * CAMERA_PARAM_COUNT :].reshape((n_points, 3))
 
@@ -169,12 +80,12 @@ def get_sparsity_pattern(n_cameras, n_points, camera_indices, point_indices):
 
 
 def bundle_adjust(camera_array: CameraArray, points_csv_path: Path):
-# Original example taken from https://scipy-cookbook.readthedocs.io/items/bundle_adjustment.html
+    # Original example taken from https://scipy-cookbook.readthedocs.io/items/bundle_adjustment.html
+
+    # MAC: start here tomorrow. You need to figure out how to cull the points
+    # based on reproj error and rerun bundle adjustment
 
     camera_params = get_camera_params(camera_array)
-
-    points_2d_df = get_points_2d_df(points_csv_path)
-    points_3d_df = get_points_3d_df(points_csv_path)
 
     (
         camera_indices,
@@ -182,7 +93,7 @@ def bundle_adjust(camera_array: CameraArray, points_csv_path: Path):
         points_2d,
         points_3d,
         n_points,
-    ) = get_bundle_adjust_params(points_2d_df, points_3d_df)
+    ) = get_2d_3d_points(points_csv_path)
 
     n_cameras = camera_params.shape[0]
     n_points = points_3d.shape[0]
